server: Don't truncate impact time sent in update
Impact time was cast to short when creating the update;
now impact time has the correct precision

+++ src/re/neutrino/java_tanks/server/Game.java
@@ -332,7 +332,7 @@ public class Game {
 	private MapPosition shotWithoutDamage(Player player, Shot shot) {
 		Impact i = map.shotWithoutDamage(player, shot);
 
-	    allAddUpdate(new ShotImpactUpdate((short)i.getTime()));
+	    allAddUpdate(new ShotImpactUpdate(i.getTime()));
 	    return i.getPos();
 	}
